Did some file moving to make the semi-automated testing simpler. Added test files
