+++ src/main/java/duke/ByeCommand.java
@@ -1,6 +1,4 @@
-package duke.commands;
-
-import duke.Duke;
+package duke;
 
 public class ByeCommand extends Command{
