+++ src/main/java/duke/ListCommand.java
@@ -1,6 +1,4 @@
-package duke.commands;
-
-import duke.Duke;
+package duke;
 
 public class ListCommand extends Command {
