+++ src/main/java/duke/Main.java
@@ -1,0 +1,8 @@
+package duke;
+
+public class Main {
+    public static void main(String args[]) {
+        Duke duke = new Duke();
+        duke.start();
+    }
+}
