+++ src/main/java/duke/TaskList.java
@@ -1,7 +1,5 @@
 package duke;
 
-import duke.tasks.Task;
-
 import java.util.ArrayList;
 
 public class TaskList {
