+++ text-ui-test/ACTUAL.txt
@@ -1,0 +1,40 @@
+	____________________________________________________________
+		Hello! I'm duke.Duke
+		What can I do for you?
+	____________________________________________________________
+	____________________________________________________________
+		Got it. I've added this task:
+			[T][✘] one two
+		Now you have 1 tasks in the list.
+	____________________________________________________________
+	____________________________________________________________
+		Got it. I've added this task:
+			[E][✘] three four (at: five six)
+		Now you have 2 tasks in the list.
+	____________________________________________________________
+	____________________________________________________________
+		Got it. I've added this task:
+			[D][✘] seven eight (at: nine ten)
+		Now you have 3 tasks in the list.
+	____________________________________________________________
+	____________________________________________________________
+		1.[T][✘] one two
+		2.[E][✘] three four (at: five six)
+		3.[D][✘] seven eight (at: nine ten)
+	____________________________________________________________
+	____________________________________________________________
+		Nice! I've marked this task as done:
+			[T][✓] one two
+	____________________________________________________________
+	____________________________________________________________
+		Nice! I've marked this task as done:
+			[E][✓] three four (at: five six)
+	____________________________________________________________
+	____________________________________________________________
+		1.[T][✓] one two
+		2.[E][✓] three four (at: five six)
+		3.[D][✘] seven eight (at: nine ten)
+	____________________________________________________________
+	____________________________________________________________
+		Bye. Hope to see you again soon!
+	____________________________________________________________
