+++ text-ui-test/runtest.sh
@@ -1,0 +1,34 @@
+#!/usr/bin/env bash
+
+# create bin directory if it doesn't exist
+if [ ! -d "../bin" ]
+then
+    mkdir ../bin
+fi
+
+# delete output from previous run
+if [ -e "./ACTUAL.txt" ]
+then
+    rm ACTUAL.txt
+fi
+
+# compile the code into the bin folder, terminates if error occurred
+if ! javac -cp ../src -Xlint:none -d ../bin ../src/main/java/duke/*.java
+then
+    echo "********** BUILD FAILURE **********"
+    exit 1
+fi
+
+# run the program, feed commands from input.txt file and redirect the output to the ACTUAL.TXT
+java -classpath ../bin duke.Main < input.txt > ACTUAL.txt
+
+# compare the output to the expected output
+diff ACTUAL.txt EXPECTED.txt
+if [ $? -eq 0 ]
+then
+    echo "Test result: PASSED"
+    exit 0
+else
+    echo "Test result: FAILED"
+    exit 1
+fi
